+++ lib/learning-objectives.js
